Moving profile domain object from articles to profiles routes
