+++ src/main/kotlin/io/github/nomisrev/routes/profiles.kt
@@ -1,0 +1,11 @@
+package io.github.nomisrev.routes
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Profile(
+    val username: String,
+    val bio: String,
+    val image: String,
+    val following: Boolean
+)
